Map Progress
Progress of the map so far. Photo 1 shows the finished borders and photo
2 shows the relative placement of realms. A map of Termina has been
included to give MM fans a way to play in a place that looks like
Termina, but plays differently. This "parallel" of Termina (it ends) I
would like to be a realm called Comiena (it begins), with a Hylian some
differences from the races that live in Termina. This can also be the
origin points for an Ikana otherworld invasion event.